refactor: extract to a new files actors `WalletFullScanScriptInspector` and `WalletSyncScriptInspector`
