+++ BDKSwiftExampleWallet/View Model/WalletViewModel.swift
@@ -163,45 +163,3 @@ class WalletViewModel {
         }
     }
 }
-
-actor WalletSyncScriptInspector: @preconcurrency SyncScriptInspector {
-    private let updateProgress: @Sendable (UInt64, UInt64) -> Void
-    private var inspectedCount: UInt64 = 0
-    private var totalCount: UInt64 = 0
-
-    init(updateProgress: @escaping @Sendable (UInt64, UInt64) -> Void) {
-        self.updateProgress = updateProgress
-    }
-
-    func inspect(script: Script, total: UInt64) {
-        totalCount = total
-        inspectedCount += 1
-
-        let delay: TimeInterval =
-            if total <= 5 {
-                0.2
-            } else if total < 10 {
-                0.15
-            } else if total < 20 {
-                0.1
-            } else {
-                0
-            }
-        Thread.sleep(forTimeInterval: delay)
-        updateProgress(inspectedCount, totalCount)
-    }
-}
-
-actor WalletFullScanScriptInspector: @preconcurrency FullScanScriptInspector {
-    private let updateProgress: @Sendable (UInt64) -> Void
-    private var inspectedCount: UInt64 = 0
-
-    init(updateProgress: @escaping @Sendable (UInt64) -> Void) {
-        self.updateProgress = updateProgress
-    }
-
-    func inspect(keychain: KeychainKind, index: UInt32, script: Script) {
-        inspectedCount += 1
-        updateProgress(inspectedCount)
-    }
-}
